Rename various form helpers
